[warning] conversion to int
fixes following warning

Implicit conversion loses integer precision: 'NSUInteger' (aka 'unsigned long') to 'int32_t' (aka 'int')

accept NumberofTouch will be in the range for int32_t (+/-2147483647)

+++ xbmc/platform/darwin/ios/XBMCController.mm
@@ -517,7 +517,7 @@ public:
       CGenericTouchActionHandler::GetInstance().OnSwipe(direction,
                                                 0.0, 0.0,
                                                 point.x, point.y, 0, 0,
-                                                [sender numberOfTouches]);
+                                                static_cast<int32_t>([sender numberOfTouches]));
     }
   }
 }
@@ -532,7 +532,7 @@ public:
     point.x *= screenScale;
     point.y *= screenScale;
     //NSLog(@"%s singleTap", __PRETTY_FUNCTION__);
-    CGenericTouchActionHandler::GetInstance().OnTap((float)point.x, (float)point.y, [sender numberOfTouches]);
+    CGenericTouchActionHandler::GetInstance().OnTap((float)point.x, (float)point.y, static_cast<int32_t>([sender numberOfTouches]));
   }
 }
 //--------------------------------------------------------------
